consensus: start test refactor
